backend/requirements.txt
Django==3.2.5
django-extensions==3.2.1
django-countries==7.2.1
django-cors-headers==3.2.1
djangorestframework==3.12.0
ipython==8.7.0
jupyter==1.0.0
psycopg2-binary
sentry-sdk